feat: 결과 페이지 요약완
